Refactor Sentry, split code between files depending on usage (browser/server/universal)
