+++ src/app/components/ServerPageBootstrap.tsx
@@ -1,7 +1,7 @@
 import { MultiversalPageProps } from '@/layouts/core/types/MultiversalPageProps';
 import { OnlyServerPageProps } from '@/layouts/core/types/OnlyServerPageProps';
 import { createLogger } from '@/modules/core/logging/logger';
-import { configureSentryUserMetadata } from '@/modules/core/sentry/sentry';
+import { configureSentryUserMetadata } from '@/modules/core/sentry/universal';
 import { userSessionContext } from '@/modules/core/userSession/userSessionContext';
 import * as Sentry from '@sentry/node';
 import React from 'react';
