+++ src/layouts/core/components/CoreLayout.tsx
@@ -3,7 +3,7 @@ import { GenericObject } from '@/modules/core/data/types/GenericObject';
 import DefaultErrorLayout from '@/modules/core/errorHandling/DefaultErrorLayout';
 import { createLogger } from '@/modules/core/logging/logger';
 import PreviewModeBanner from '@/modules/core/previewMode/components/PreviewModeBanner';
-import Sentry from '@/modules/core/sentry/sentry';
+import Sentry from '@/modules/core/sentry/init';
 import ErrorPage from '@/pages/_error';
 import {
   Amplitude,
