+++ src/pages/api/preview.ts
@@ -5,7 +5,8 @@ import {
 } from '@/modules/core/amplitude/events';
 import { filterExternalAbsoluteUrl } from '@/modules/core/js/url';
 import { createLogger } from '@/modules/core/logging/logger';
-import Sentry, { configureReq } from '@/modules/core/sentry/sentry';
+import Sentry from '@/modules/core/sentry/init';
+import { configureReq } from '@/modules/core/sentry/server';
 import appendQueryParameter from 'append-query';
 import {
   NextApiRequest,
