+++ src/pages/api/status.ts
@@ -4,7 +4,8 @@ import {
   AMPLITUDE_EVENTS,
 } from '@/modules/core/amplitude/events';
 import { createLogger } from '@/modules/core/logging/logger';
-import Sentry, { configureReq } from '@/modules/core/sentry/sentry';
+import Sentry from '@/modules/core/sentry/init';
+import { configureReq } from '@/modules/core/sentry/server';
 import {
   NextApiRequest,
   NextApiResponse,
